# taskmanagement/openapi/main.yml
openapi: 3.0.3
info:
  title: Rest API
  description: All Item Rest API End Points
  version: V1-2024-08
servers:
  - url: http://localhost:8080
tags:
  - name:  Admin
  - name:  Task
paths:
  /admin:
    post:
      tags:
        - Admin
      summary: create a new Admin
      description: add a new Admin in app
      operationId: createAdmin
      requestBody:
        description: Admin data required to create a new Admin
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/adminDto'
      responses:
        '200':
          description: Admin created
          content:
            application/json:
              schema:
                type: string
        '409' :
          description: Admin Already Exist

  /admin/task:
    post:
      tags:
        - Task
      summary: create task
      description: create task
      operationId: createTask
      requestBody:
        description: Task data to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/taskDto'
      responses:
        '200':
          description: the add task was successful
          content:
            application/json:
              schema:
                type: string


  /admin/{adminId}:
    get:
      tags:
        - Admin
      summary: get Admin by id
      description: get Admin By id
      operationId: fetchAdminById
      parameters:
        - name: adminId
          in: path
          description: Id of the Admin to fetch
          required: true
          schema:
            type: string
      responses:
        '200' :
          description: Admin to fetch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/adminDto'

    put:
      tags:
        - Admin
      summary: Update Admin by id
      description: Update Admin by id
      operationId: updateAdminById
      parameters:
        - name: adminId
          in: path
          description: Id of the admin to upDate
          required: true
          schema:
            type: string
      requestBody:
        description: Admin to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/adminDto'
      responses:
        '200' :
          description: the update was successful
          content:
            application/json:
              schema:
                type: string

    delete:
      tags:
         - Admin
      summary: Delete Admin by Id
      description: Delete Admin by Id
      operationId: deleteAdminById
      parameters:
        - name: adminId
          in: path
          description: Id admin to delete
          required: true
          schema:
            type: string
      responses:
        '200' :
          description: the delete was successful
          content:
            application/json:
              schema:
                type: string


components:
  schemas:
    adminDto:
      type: object
      properties:
        firstName:
          type: string
          description: firstName of the Admin
        lastName:
          type: string
          description: lastName of the Admin
        email:
          type: string
          description: email of the Admin

    taskDto:
      type: object
      properties:
        title:
          description: title of the task
          type: string
        description:
          type: string
          description: description of the task
        priority:
          type: object
          description: priority of the task
        status:
          type: object
          description: status of the task
        dueDate:
          type: string
          description: the creation date of the task
        adminId:
          type: integer
          description: Id of the admin who created the task